[HER-16] Rename user to index
